luajit/x86: sign-extend disp32 in jit_base extraction (x86asm gives it unsigned)
x86asm.Decode reports a negative disp32 as its unsigned 32-bit value in the
int64 Disp field (0xFFFFF1E0, not -3616), so the g2dispatch+disp candidate
overflowed the validation range and the extractor fell back to cur_L+8 (mem_L).
Sign-extend the low 32 bits. Verified against the tarantool binary: g2jitbase
now extracts 0x1f0 (jit_base) instead of 0x1e8 (mem_L).

+++ interpreter/luajit/extractor_x86.go
@@ -242,8 +242,10 @@ func (x *x86Extractor) findG2JitBaseFromExitHandler(
 			mem, ok0 := i.Args[0].(x86asm.Mem)
 			imm, ok1 := i.Args[1].(x86asm.Imm)
 			if ok0 && ok1 && imm == 0 && mem.Base == x86asm.R14 && mem.Index == 0 {
-				// Address wraparound handles the negative DISPATCH-relative displacement.
-				cand := g2dispatch + libpf.Address(mem.Disp)
+				// x86asm reports a disp32 as its unsigned value in an int64
+				// (e.g. 0xFFFFF1E0 rather than -3616), so sign-extend the low 32
+				// bits before adding. Address wraparound then yields g2dispatch+disp.
+				cand := g2dispatch + libpf.Address(int64(int32(mem.Disp)))
 				if cand > curLOffset && cand <= curLOffset+0x18 {
 					return cand, nil
 				}
